Added casting state to player state machine.
Player will now enter a casting state (initial state for casting transitions) when pressing the cast key (currently 'Z' on the keyboard). Currently, letting go of the 'Z' key will return the player to the idling state. In the future, this will lead towards the launch and animation of the abilities. To cast, all the former player states will now inherit from a Castable class.